refactor: add some styles
