+++ src/components/layout/layout.scss
@@ -1,5 +1,5 @@
 .container {
-  max-width: 1440px;
+  max-width: 1280px;
   margin: auto;
   padding: 20px;
 }
